api/requirements.txt
Django==2.2.12
psycopg2>=2.7,<3.0
djangorestframework==3.11.0